fix: reject beacon state requests while syncing to avoid regen wedge
A far-behind node polled for a state via REST can trigger a regen that walks back past the
block-root window (`SLOTS_PER_HISTORICAL_ROOT`) and wedges (blocks the main thread, 0 blocks
imported). Guard the shared `getStateResponseWithRegen` entry point with the existing
`notWhileSyncing` check so such requests return 503 (`NodeIsSyncing`) while the node is behind,
matching the validator duties endpoints.

- Extract `notWhileSyncing` + `SYNC_TOLERANCE_EPOCHS` from the validator API into the shared
  `api/impl/utils.ts` and reuse them (no behavior change for validator endpoints).
- Thread `sync` into `getStateResponseWithRegen` and its callers (beacon state, proof, debug,
  lodestar, validator) so the single choke point covers all REST state-serving endpoints.
- Test: `getStateResponseWithRegen` throws `NodeIsSyncing` while syncing.

Alternative to #9634 (per @nflaig review) — simpler, and covers all REST state endpoints rather
than only the far-slot regen case.
